only use skeletons for data-dependent components, not section headers
